backend/requirements.txt
aioboto3==15.1.0
aiobotocore==2.24.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
audioop-lts==0.2.1
av==13.1.0
beautifulsoup4==4.13.4
boto3==1.39.11
botocore==1.39.11
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
clerk-backend-api==3.1.11
click==8.2.1
cloup==3.0.7
cryptography==45.0.5
cython==3.1.2
decorator==5.2.1
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
filetype==1.2.0
frozenlist==1.7.0
glcontext==3.0.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
isosurfaces==0.1.2
jinja2==3.1.6
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.72
langchain-google-genai==2.1.8
langchain-text-splitters==0.3.9
langsmith==0.4.8
mako==1.3.10
manim==0.19.0
manimpango==0.6.0
mapbox-earcut==1.0.3
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
moderngl==5.12.0
moderngl-window==3.1.1
multidict==6.6.4
mypy==1.17.0
mypy-extensions==1.1.0
networkx==3.5
numpy==2.3.2
orjson==3.11.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycairo==1.28.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydub==0.25.1
pyglet==2.1.6
pyglm==2.8.2
pygments==2.19.2
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pyyaml==6.0.2
redis==6.4.0
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.14.8
rignore==0.6.4
rsa==4.9.1
s3transfer==0.13.1
scipy==1.16.1
screeninfo==0.8.1
sentry-sdk==2.33.2
shellingham==1.5.4
six==1.17.0
skia-pathops==0.8.0.post2
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.41
srt==3.5.3
starlette==0.47.2
svgelements==1.9.6
tenacity==9.1.2
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1
zstandard==0.23.0
